+++ frontend/src/components/prompts/Share.vue
@@ -75,13 +75,29 @@
         </i>
       </p>
       <input class="input" type="password" autocomplete="new-password" v-model.trim="password" />
+      <p>
+        {{ $t("share.shareType") }}
+        <i
+          class="no-select material-symbols-outlined tooltip-info-icon"
+          @mouseenter="showTooltip($event, $t('share.shareTypeDescription'))"
+          @mouseleave="hideTooltip"
+        >
+          help
+        </i>
+      </p>
+      <select class="input" v-model="shareType">
+        <option value="normal">{{ $t("share.normalShare") }}</option>
+        <option value="upload">{{ $t("share.uploadShare") }}</option>
+      </select>
+
       <div class="settings-items" style="margin-top: 0.5em;">
         <!--
         <ToggleSwitch class="item" v-model="allowEdit" :name="'Allow modifications'" />
         <ToggleSwitch class="item" v-model="allowUpload" :name="'Allow uploading'" />
         -->
-        <ToggleSwitch class="item" v-model="disableFileViewer" :name="$t('share.disableFileViewer')" />
+        <ToggleSwitch v-if="shareType === 'normal'" class="item" v-model="disableFileViewer" :name="$t('share.disableFileViewer')" />
         <ToggleSwitch
+          v-if="shareType === 'normal'"
           class="item"
           v-model="quickDownload"
           :name="$t('profileSettings.showQuickDownload')"
@@ -90,8 +106,8 @@
         <ToggleSwitch class="item" v-model="disableAnonymous" :name="$t('share.disableAnonymous')" :description="$t('share.disableAnonymousDescription')" />
         <ToggleSwitch class="item" v-model="enableAllowedUsernames" :name="$t('share.enableAllowedUsernames')" :description="$t('share.enableAllowedUsernamesDescription')" />
 
-        <ToggleSwitch v-if="onlyOfficeAvailable" class="item" v-model="enableOnlyOffice" :name="$t('share.enableOnlyOffice')" :description="$t('share.enableOnlyOfficeDescription')" />
-        <ToggleSwitch v-if="onlyOfficeAvailable" class="item" v-model="enableOnlyOfficeEditing" :name="$t('share.enableOnlyOfficeEditing')" :description="$t('share.enableOnlyOfficeEditingDescription')" />
+        <ToggleSwitch v-if="shareType === 'normal' && onlyOfficeAvailable" class="item" v-model="enableOnlyOffice" :name="$t('share.enableOnlyOffice')" :description="$t('share.enableOnlyOfficeDescription')" />
+        <ToggleSwitch v-if="shareType === 'normal' && onlyOfficeAvailable" class="item" v-model="enableOnlyOfficeEditing" :name="$t('share.enableOnlyOfficeEditing')" :description="$t('share.enableOnlyOfficeEditingDescription')" />
         <p>
           {{ $t("share.enforceDarkLightMode") }}
           <i
@@ -130,53 +146,57 @@
             </option>
           </select>
         </div>
-        <p>
-          {{ $t("share.defaultViewMode") }}
-          <i
-            class="no-select material-symbols-outlined tooltip-info-icon"
-            @mouseenter="showTooltip($event, $t('share.defaultViewModeDescription'))"
-            @mouseleave="hideTooltip"
-          >
-            help
-          </i>
-        </p>
-        <select class="input" v-model="viewMode">
-          <option value="normal">{{ $t("buttons.normalView") }}</option>
-          <option value="list">{{ $t("buttons.listView") }}</option>
-          <option value="compact">{{ $t("buttons.compactView") }}</option>
-          <option value="gallery">{{ $t("buttons.galleryView") }}</option>
-        </select>
+        <div v-if="shareType === 'normal'">
+          <p>
+            {{ $t("share.defaultViewMode") }}
+            <i
+              class="no-select material-symbols-outlined tooltip-info-icon"
+              @mouseenter="showTooltip($event, $t('share.defaultViewModeDescription'))"
+              @mouseleave="hideTooltip"
+            >
+              help
+            </i>
+          </p>
+          <select class="input" v-model="viewMode">
+            <option value="normal">{{ $t("buttons.normalView") }}</option>
+            <option value="list">{{ $t("buttons.listView") }}</option>
+            <option value="compact">{{ $t("buttons.compactView") }}</option>
+            <option value="gallery">{{ $t("buttons.galleryView") }}</option>
+          </select>
+        </div>
       <SettingsItem :title="$t('buttons.showMore')" :collapsable="true" :start-collapsed="true">
         <div class="settings-items">
           <ToggleSwitch class="item" v-model="keepAfterExpiration" :name="$t('share.keepAfterExpiration')" :description="$t('share.keepAfterExpirationDescription')" />
-          <ToggleSwitch class="item" v-model="disableThumbnails" :name="$t('share.disableThumbnails')" :description="$t('share.disableThumbnailsDescription')" />
+          <ToggleSwitch v-if="shareType === 'normal'" class="item" v-model="disableThumbnails" :name="$t('share.disableThumbnails')" :description="$t('share.disableThumbnailsDescription')" />
           <ToggleSwitch class="item" v-model="disableNavButtons" :name="$t('share.hideNavButtons')" :description="$t('share.hideNavButtonsDescription')" />
           <ToggleSwitch class="item" v-model="disableShareCard" :name="$t('share.disableShareCard')" :description="$t('share.disableShareCardDescription')" />
-          <ToggleSwitch class="item" v-model="disableSidebar" :name="$t('share.disableSidebar')" :description="$t('share.disableSidebarDescription')" />
+          <ToggleSwitch v-if="shareType === 'normal'" class="item" v-model="disableSidebar" :name="$t('share.disableSidebar')" :description="$t('share.disableSidebarDescription')" />
         </div>
 
-        <p>
-          {{ $t("prompts.downloadsLimit") }}
-          <i
-            class="no-select material-symbols-outlined tooltip-info-icon"
-            @mouseenter="showTooltip($event, $t('share.downloadsLimitDescription'))"
-            @mouseleave="hideTooltip"
-          >
-            help
-          </i>
-        </p>
-        <input class="input" type="number" min="0" v-model.number="downloadsLimit" />
-        <p>
-          {{ $t("prompts.maxBandwidth") }}
-          <i
-            class="no-select material-symbols-outlined tooltip-info-icon"
-            @mouseenter="showTooltip($event, $t('share.maxBandwidthDescription'))"
-            @mouseleave="hideTooltip"
-          >
-            help
-          </i>
-        </p>
-        <input class="input" type="number" min="0" v-model.number="maxBandwidth" />
+        <div v-if="shareType === 'normal'">
+          <p>
+            {{ $t("prompts.downloadsLimit") }}
+            <i
+              class="no-select material-symbols-outlined tooltip-info-icon"
+              @mouseenter="showTooltip($event, $t('share.downloadsLimitDescription'))"
+              @mouseleave="hideTooltip"
+            >
+              help
+            </i>
+          </p>
+          <input class="input" type="number" min="0" v-model.number="downloadsLimit" />
+          <p>
+            {{ $t("prompts.maxBandwidth") }}
+            <i
+              class="no-select material-symbols-outlined tooltip-info-icon"
+              @mouseenter="showTooltip($event, $t('share.maxBandwidthDescription'))"
+              @mouseleave="hideTooltip"
+            >
+              help
+            </i>
+          </p>
+          <input class="input" type="number" min="0" v-model.number="maxBandwidth" />
+        </div>
 
 
         <p>
@@ -335,6 +355,7 @@ export default {
       viewMode: "normal",
       enableOnlyOffice: false,
       enableOnlyOfficeEditing: false,
+      shareType: "normal",
       //viewMode: "normal",
     };
   },
@@ -417,6 +438,7 @@ export default {
           this.viewMode = this.link.viewMode || "normal";
           this.enableOnlyOffice = this.link.enableOnlyOffice || false;
           this.enableOnlyOfficeEditing = this.link.enableOnlyOfficeEditing || false;
+          this.shareType = this.link.shareType || "normal";
           //this.viewMode = this.link.viewMode || "normal";
         }
       },
@@ -499,6 +521,7 @@ export default {
           viewMode: this.viewMode,
           enableOnlyOffice: this.enableOnlyOffice,
           enableOnlyOfficeEditing: this.enableOnlyOfficeEditing,
+          shareType: this.shareType,
         };
         if (this.isEditMode) {
           payload.hash = this.link.hash;